+++ src/backend/InvenTree/InvenTree/helpers_model.py
@@ -3,12 +3,10 @@
 import io
 import ipaddress
 import socket
-from decimal import Decimal
 from typing import Optional, cast
 from urllib.parse import urljoin, urlparse
 
 from django.conf import settings
-from django.core.exceptions import ValidationError
 from django.core.validators import URLValidator
 from django.db.utils import OperationalError, ProgrammingError
 from django.utils.translation import gettext_lazy as _
@@ -16,8 +14,6 @@ from django.utils.translation import gettext_lazy as _
 import requests
 import requests.exceptions
 import structlog
-from djmoney.contrib.exchange.models import convert_money
-from djmoney.money import Money
 from PIL import Image
 
 from common.notifications import (
@@ -31,7 +27,6 @@ from InvenTree.cache import (
     get_session_cache,
     set_session_cache,
 )
-from InvenTree.format import format_money
 from InvenTree.ready import ignore_ready_warning
 
 logger = structlog.get_logger('inventree')
@@ -250,90 +245,6 @@ def download_image_from_url(
         raise TypeError(_('Supplied URL is not a valid image file'))
 
     return img
-
-
-def render_currency(
-    money: Money,
-    decimal_places: Optional[int] = None,
-    currency: Optional[str] = None,
-    multiplier: Optional[Decimal] = None,
-    min_decimal_places: Optional[int] = None,
-    max_decimal_places: Optional[int] = None,
-    include_symbol: bool = True,
-    locale: Optional[str] = None,
-):
-    """Render a currency / Money object to a formatted string (e.g. for reports).
-
-    Arguments:
-        money: The Money instance to be rendered
-        decimal_places: The number of decimal places to render to. If unspecified, uses the PRICING_DECIMAL_PLACES setting.
-        currency: Optionally convert to the specified currency
-        multiplier: An optional multiplier to apply to the money amount before rendering
-        min_decimal_places: The minimum number of decimal places to render to. If unspecified, uses the PRICING_DECIMAL_PLACES_MIN setting.
-        max_decimal_places: The maximum number of decimal places to render to. If unspecified, uses the PRICING_DECIMAL_PLACES setting.
-        include_symbol: If True, include the currency symbol in the output
-        locale: Optional locale override (e.g. 'en-us', 'en-gb'). Controls symbol vs. code rendering. Defaults to server LANGUAGE_CODE.
-    """
-    if money in [None, '']:
-        return '-'
-
-    if type(money) is not Money:
-        # Try to convert to a Money object
-        try:
-            money = Money(
-                Decimal(str(money)),
-                currency or get_global_setting('INVENTREE_DEFAULT_CURRENCY'),
-            )
-        except Exception:
-            raise ValidationError(
-                f"render_currency: {_('Invalid money value')}: '{money}' ({type(money).__name__})"
-            )
-
-    if currency is not None:
-        # Attempt to convert to the provided currency
-        # If cannot be done, leave the original
-        try:
-            money = convert_money(money, currency)
-        except Exception:
-            pass
-
-    if multiplier is not None:
-        try:
-            money *= Decimal(str(multiplier).strip())
-        except Exception:
-            raise ValidationError(
-                f"render_currency: {_('Invalid multiplier value')}: '{multiplier}' ({type(multiplier).__name__})"
-            )
-
-    if min_decimal_places is None or not isinstance(min_decimal_places, (int, float)):
-        min_decimal_places = get_global_setting('PRICING_DECIMAL_PLACES_MIN', 0)
-
-    if max_decimal_places is None or not isinstance(max_decimal_places, (int, float)):
-        max_decimal_places = get_global_setting('PRICING_DECIMAL_PLACES', 6)
-
-    value = Decimal(str(money.amount)).normalize()
-    value = str(value)
-
-    if decimal_places is not None and isinstance(decimal_places, (int, float)):
-        # Decimal place count is provided, use it
-        pass
-    elif '.' in value:
-        # If the value has a decimal point, use the number of decimal places in the value
-        decimal_places = len(value.split('.')[-1])
-    else:
-        # No decimal point, use 2 as a default
-        decimal_places = 2
-
-    # Clip the decimal places to the specified range
-    decimal_places = max(decimal_places, min_decimal_places)
-    decimal_places = min(decimal_places, max_decimal_places)
-
-    return format_money(
-        money,
-        decimal_places=decimal_places,
-        include_symbol=include_symbol,
-        locale=locale,
-    )
 
 
 @ignore_ready_warning
